refactor: move target services out of assembly_service
Extract the ten target category/value functions into a new
service_layer/target_service.py, verbatim and with no logic changes.
assembly_service.py drops from 1228 to 834 lines.

No re-export shim: callers import from the new module directly.
