+++ backend/src/opendlp/entrypoints/blueprints/targets_legacy.py
@@ -10,15 +10,7 @@ from flask_login import current_user, login_required
 
 from opendlp import bootstrap
 from opendlp.service_layer.assembly_service import (
-    add_target_value,
-    create_target_category,
-    delete_target_category,
-    delete_target_value,
     get_assembly_with_permissions,
-    get_targets_for_assembly,
-    import_targets_from_csv,
-    update_target_category,
-    update_target_value,
 )
 from opendlp.service_layer.constants import MAX_DISTINCT_VALUES_FOR_AUTO_ADD
 from opendlp.service_layer.exceptions import InsufficientPermissions, InvalidSelection, NotFoundError
@@ -30,6 +22,16 @@ from opendlp.service_layer.target_respondent_helpers import (
     get_assembly_respondent_attribute_columns,
     get_column_distinct_counts,
     get_respondent_counts_for_category,
+)
+from opendlp.service_layer.target_service import (
+    add_target_value,
+    create_target_category,
+    delete_target_category,
+    delete_target_value,
+    get_targets_for_assembly,
+    import_targets_from_csv,
+    update_target_category,
+    update_target_value,
 )
 from opendlp.translations import gettext as _
