+++ backend/tests/e2e/test_targets_legacy_pages.py
@@ -8,7 +8,7 @@ from opendlp.domain.respondents import Respondent
 from opendlp.domain.selection_settings import SelectionSettings
 from opendlp.domain.users import UserAssemblyRole
 from opendlp.domain.value_objects import AssemblyRole
-from opendlp.service_layer.assembly_service import (
+from opendlp.service_layer.target_service import (
     add_target_value,
     create_target_category,
     import_targets_from_csv,
